Module Generalised_signatures.Hashtbl

module type Hashtbl_generalised = sig ... end

The most general hashtable signature, supporting the four specialisations below.

module type Poly_hash = sig ... end

A hashtable with polymorphic keys and values, such as Stdlib.Hashtbl.

module type Mono_hash = sig ... end

A hashtable with fixed key type and hashing function, such as Stdlib.Hashtbl.S.

module type Persistent = sig ... end

A hashtable in which bindings are serialised to disk, such as Index.

module type Generic = sig ... end

A heterogeneous hashtable in which the type of a key determines the type of the value that it points to, such as Core_kernel.Univ_map.