Module type Hashtbl.Hashtbl_generalised

The most general hashtable signature, supporting the four specialisations below.

We have three types (t, key and value) and three type variables:

'k/'v allow the hashtable to determine key/value types;

'a is carried from keys to corresponding values, allowing the key to determine the types of values.

type ('k, 'v) t

type ('k, 'a) key

type ('v, 'a) value

val create : int -> (_, _) t

val replace : ('k, 'v) t -> ('k, 'a) key -> ('v, 'a) value -> unit

val remove : ('k, _) t -> ('k, _) key -> unit

val find_opt : ('k, 'v) t -> ('k, 'a) key -> ('v, 'a) value option